avoid calling finalizers on types that don't require it (part 1)

+++ common/Animations/Keyframe.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 
 /// <summary> Represents a keyframe with a time and value. </summary>
 /// <typeparam name="TValue"> A type that represents the value of this keyframe. </typeparam>
-public readonly struct Keyframe<TValue>(float time, TValue value, Func<double, double> easing, bool until)
+public readonly struct Keyframe<TValue>(float time, TValue value, Func<float, float> easing, bool until)
     : IEquatable<Keyframe<TValue>>, IComparer<Keyframe<TValue>>
 {
     ///<summary> Time of this keyframe. </summary>
@@ -15,7 +15,7 @@ public readonly struct Keyframe<TValue>(float time, TValue value, Func<double, d
     public readonly TValue Value = value;
 
     /// <summary> <see cref="EasingFunctions"/> easing of this keyframe. </summary>
-    public readonly Func<double, double> Ease = easing ?? EasingFunctions.Linear;
+    public readonly Func<float, float> Ease = easing ?? EasingFunctions.Linear;
 
     /// <summary> Reserved for <see cref="Storyboarding.Util.CommandGenerator"/>. </summary>
     internal readonly bool Until = until;
@@ -24,7 +24,7 @@ public readonly struct Keyframe<TValue>(float time, TValue value, Func<double, d
     /// <param name="time"> Time of the keyframe. </param>
     /// <param name="value"> A value to be assigned to the keyframe. </param>
     /// <param name="easing"> <see cref="EasingFunctions"/> easing to be assigned. </param>
-    public Keyframe(float time, TValue value = default, Func<double, double> easing = null) : this(time, value, easing, false) { }
+    public Keyframe(float time, TValue value = default, Func<float, float> easing = null) : this(time, value, easing, false) { }
 
     internal static readonly Comparer<Keyframe<TValue>> Comparer =
         Comparer<Keyframe<TValue>>.Create((x, y) => Math.Sign(x.Time - y.Time));
